fail deployment if a release exceeds 50 artifacts
Following commit 0155cabbb7b22494b3d5da1425059387d8236cf2, the number of
artifacts produced during packaging is down to 40. We expect this count to rise
as the CMake build system continues to be worked on, operating systems move in
and out of support cycles, and the Cygwin/Msys builds are tweaked. This change
introduced a safe guard to fail deployment if the number of artifacts grows to
exceed 50 again.

Related to #2147.