jpeg_size: remove use of 'junk' variable
This code was reading big endian integers into this variable that was then
ignored as a way of skipping fields. It is simpler and more efficient to just
fast forward the file handle over those bytes. This assumes the file handle is
to something seekable (e.g. not a FIFO), but the handle was already seeked
elsewhere in this function, so this assumption already existed.